Fix session vote state stuck
When changing session participants to observers their score is removed. This means you can end up in a state where votes are visible but now nobody has a score submitted and so further voting is locked.

An erroneous condition on allowing clearing of votes meant that the clear button was also locked and so the session was stuck in a state of having no votes but also being unable to reset 🙃